separation_1

+++ source/handlers/user/configs_menu/delete_config.py
@@ -1,10 +1,11 @@
 from aiogram import types
 from aiogram.dispatcher import FSMContext
+import aiohttp
 
 from source.keyboard import inline
 from source.middlewares import rate_limit
 from source.utils import localizer
-from source.utils.xray import xray_config
+from loader import db_manager
 
 
 @rate_limit(limit=1)
@@ -23,19 +24,52 @@ async def confirm_delete_config(call: types.CallbackQuery, state: FSMContext):
     )
 
 
+# async def delete_config(call: types.CallbackQuery, state: FSMContext):
+#     config_uuid = call.data.split("_")[-1]
+#     await xray_config.disconnect_user_by_uuid(uuid=config_uuid)
+#     await db_manager.delete_one_vpn_config_by_uuid(uuid=uuid)
+#     await call.message.answer(
+#         text=localizer.get_user_localized_text(
+#             user_language_code=call.from_user.language_code,
+#             text_localization=localizer.message.config_deleted,
+#         ),
+#         parse_mode=types.ParseMode.HTML,
+#         reply_markup=await inline.insert_button_back_to_main_menu(
+#             language_code=call.from_user.language_code
+#         ),
+#     )
+#     await call.answer()
+#     await call.message.delete()
+#     await state.finish()
+
+
 async def delete_config(call: types.CallbackQuery, state: FSMContext):
-    config_uuid = call.data.split("_")[-1]
-    await xray_config.disconnect_user_by_uuid(uuid=config_uuid)
-    await call.message.answer(
-        text=localizer.get_user_localized_text(
-            user_language_code=call.from_user.language_code,
-            text_localization=localizer.message.config_deleted,
-        ),
-        parse_mode=types.ParseMode.HTML,
-        reply_markup=await inline.insert_button_back_to_main_menu(
-            language_code=call.from_user.language_code
-        ),
-    )
-    await call.answer()
-    await call.message.delete()
-    await state.finish()
+    config_uuid = call.data.split("_")[-1]  # Получаем UUID конфигурации из данных callback
+    
+    # Получаем домен для конфигурации
+    target_server = await db_manager.get_domain_by_uuid(uuid=config_uuid)
+        
+    # Формируем URL для удаления конфигурации с учётом target_server
+    url = f"http://nginxtest.vpnizator.online/delete_config/{target_server}/"
+        
+    # Используем aiohttp для отправки POST-запроса на эндпоинт
+    async with aiohttp.ClientSession() as session:
+        async with session.post(url, json={"config_uuid": config_uuid}) as response:
+            if response.status == 200:
+                # Успешно удалили конфиг на сервере, теперь удаляем его из базы данных
+                await db_manager.delete_one_vpn_config_by_uuid(uuid=config_uuid)
+                    
+                # Отправляем сообщение о том, что конфиг удален
+                await call.message.answer(
+                    text=localizer.get_user_localized_text(
+                        user_language_code=call.from_user.language_code,
+                        text_localization=localizer.message.config_deleted,
+                    ),
+                    parse_mode=types.ParseMode.HTML,
+                    reply_markup=await inline.insert_button_back_to_main_menu(
+                        language_code=call.from_user.language_code
+                    ),
+                )
+                await call.answer()
+                await call.message.delete()
+                await state.finish()
